requirements.txt
certifi==2022.6.15
click==8.1.3
dnspython==2.2.1
Flask==2.2.0
Flask-WTF==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
python-dotenv==0.20.0
Werkzeug==2.2.1
WTForms==3.0.1
